size_t: update gmath lib
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@52045 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/gmath/ialloc.c
@@ -61,7 +61,7 @@ int **G_alloc_imatrix(int rows, int cols)
     int i;
 
     m = (int **)G_calloc(rows, sizeof(int *));
-    m[0] = (int *)G_calloc(rows * cols, sizeof(int));
+    m[0] = (int *)G_calloc((size_t) rows * cols, sizeof(int));
     for (i = 1; i < rows; i++)
 	m[i] = m[i - 1] + cols;
